+++ config/webpack/plugins/environment.ts
@@ -4,7 +4,8 @@ const {
   PORT,
   GOOGLE_ANALYTICS_MEASUREMENT_ID,
   GOOGLE_API_KEY,
-  GOOGLE_CLIENT_ID
+  GOOGLE_CLIENT_ID,
+  GOOGLE_GAE_SERVICE_URL
 } = process.env;
 
 const plugin = new EnvironmentPlugin({
@@ -13,7 +14,8 @@ const plugin = new EnvironmentPlugin({
   PORT,
   GOOGLE_ANALYTICS_MEASUREMENT_ID,
   GOOGLE_API_KEY,
-  GOOGLE_CLIENT_ID
+  GOOGLE_CLIENT_ID,
+  GOOGLE_GAE_SERVICE_URL
 });
 
 export default plugin;